feat: initial commit
Completely forgot to initialize git before starting the project. So some features are already implemented, but the git history is empty. This commit will serve as the initial commit.

Just a breakdown of what I've done so far:
- Created a Django project named `shopperly`.
- Created three apps: `accounts`, `pages`, and `shop`.
- Implemented user authentication in the `accounts` app.
- Created a basic homepage in the `pages` app.
- Developed a simple product catalog in the `shop` app.
- Added basic templates for rendering HTML pages.

I have set up the database and created initial migrations for the models in each app. The project is structured to allow for easy expansion in the future.

And also, the UI is very basic to put it nicely; to be frank it's quite ugly. But it serves the purpose of demonstrating the functionality of the apps. I plan to improve the UI in future commits, but for now, I just wanted to get the basic functionality working.